JP-1122 - Revise persistence Device ... Moved LogBooks

+++ comu/com.energyict.mdc.protocols/src/main/java/com/energyict/protocolimplv2/ace4000/ACE4000MessageExecutor.java
@@ -76,7 +76,7 @@ public class ACE4000MessageExecutor implements MessageProtocol {
             if (messageContent.contains(READ_EVENTS)) {
                 //TODO what if the offline device has no logbooks configured??
                 ReadMeterEvents readMeterEventsRequest = new ReadMeterEvents(ace4000);
-                List<CollectedLogBook> collectedLogBooks = readMeterEventsRequest.request(new LogBookIdentifierById(ace4000.getOfflineDevice().getAllOfflineLogBooks().get(0).getLogBookId()));
+                List<CollectedLogBook> collectedLogBooks = readMeterEventsRequest.request(new LogBookIdentifierById((int) ace4000.getOfflineDevice().getAllOfflineLogBooks().get(0).getLogBookId()));
                 MeterData meterData = new MeterData();
                 for (MeterProtocolEvent collectedMeterEvent : collectedLogBooks.get(0).getCollectedMeterEvents()) {
                     meterData.addMeterEvent(collectedMeterEvent);
